Allow users to specify AWS region and AMI by passing parameters to scripts

+++ scenarios/scenario-02-device-edge/aws/clean-up.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-export AWS_REGION=eu-north-1
+export AWS_REGION=$1
+if [ -z "$AWS_REGION" ]
+then
+    echo "Provide the name of the AWS region where the MLOps scenario was deployed to run the clean up script"
+    exit 1
+fi
 
 aws ec2 delete-key-pair --key-name mlops --region $AWS_REGION
 rm -f aws/mlops.pem
